fix: moved modals to components
